Slight refinement to the collisions model
Whereas previously the "winner" in a collision of two masses would be the one with the lower index of the two, the winner is now determined by the relative size of the two masses. If the mass of the two masses is the same, though, the mass with the lower index still wins.