Remove collisions
Collisions will be back, but not until an integrator with an adaptive
time step has been introduced. Problem is that without an adaptive time
step, bodies shoot through each other if the distance they travel in one
tick is greater than the sum of the radiae of the bodies that collide.