Don't calculate acceleration induced by massless particles
Kilometer sized asteroids exert close to no influence on other celestial objects, which is why it's fair game to assign a mass of zero to them. Moreover, when a mass has a mass of zero, we don't have to calculate the acceleration it induces on other celestial objects, since as per the maths, it will always be equal to zero, which saves us a great deal of computer power. Moreover, if the user increases the mass of massless object, the acceleration induced by this object on others will be taken into account, so the user still has full freedom to screw around with the universe; yay!