[Version 0.1.3][Feature] Add collisions functionality
Just an initial commit for the collisions functionality. For now collisions are set to false and won't be available in the production app. On a side note, let's start commenting and explaining the code as we work with it to make it easier for others to work with. Don't be a lazy bum!