Get rid of the loading manager for mass manifestations
Apparently, as of THREE.js r87, progress events are no longer supported, which means we cannot use the loading manager to create a loading bar that keeps track of how many percent of the resources needed for a simulation have been loaded, so let's get rid of it for the mass manifestations and revert back to our old way of handling the loading of textures, for now. There's supposed to be a work around for this issue, but I'll look into that at some later point; for now it is not a priority. Do note, however, that we keep the loading manager for the starfield texture.