[Version 0.1.6][Feature] Add dynamic star surface shader
Bless you Stefan Gustavsson and Ashima arts for writing that shader!!! So... Now we have stars with accurate colors, and their surfaces are dynamic with transient sun spots and the whole shebang. Should probably optimize this part of the code in the future, and include an option for folks with slower devices to turn off star shaders, though!