Do collisions after the particle system has iterated
What happens otherwise is that the oldMasses array will contain the celestial object that was destroyed, and when the verlet particle integrator checks for collisions when it calculates the accelerations for the oldMasses, it will detect a collision between the object that collided and the particles produced from that collision, which is not the expected behaviour. The solution to this problem is either to remove the mass that collided from the old masses array, or for no price at all, simply do the collisions after the particles have had their new state vectors generated: we choose the latter.